Allow the default image loader to be compiled out
div relies on img_loader/img_loader.h for the actual loading of the
images by default. That library was developed specifically for div, but
by no means should users be forced to use it. It is only accessed from
defaults.c so should be simple to replace with something else.

A new make flag CUSTOM_IMAGE_LOADER has been added and when set to 1,
will remove references to the default loader. div will still run, but of
course won't be able to load images any more. Users would have to
specify their own loader in this case.

This is a testament to the goal of users being able to remove anything
they don't want.