Reduce memory usage when zooming
If we zoomed to far in we'd end up creating too large of an image and
which may cause the X server to close our connection. Now we selecting a
sub image, we'll only select the parts we'll actually draw. This allowed
zooming us to zoom to the max level without issue.